Merging ART to ART branch

+++ src/common/types/tuple.cpp
@@ -10,16 +10,16 @@ using namespace std;
 TupleSerializer::TupleSerializer() : base_size(0), has_variable_columns(false) {
 }
 
-TupleSerializer::TupleSerializer(const vector<TypeId> &types_, vector<size_t> columns_) : TupleSerializer() {
+TupleSerializer::TupleSerializer(const vector<TypeId> &types_, vector<uint64_t> columns_) : TupleSerializer() {
 	Initialize(types_, columns_);
 }
 
-void TupleSerializer::Initialize(const vector<TypeId> &types_, vector<size_t> columns_) {
+void TupleSerializer::Initialize(const vector<TypeId> &types_, vector<uint64_t> columns_) {
 	types = types_;
 
 	if (columns_.size() == 0) {
 		// if _columns is not supplied we use all columns
-		for (size_t i = 0; i < types.size(); i++) {
+		for (uint64_t i = 0; i < types.size(); i++) {
 			columns.push_back(i);
 		}
 	} else {
@@ -29,7 +29,7 @@ void TupleSerializer::Initialize(const vector<TypeId> &types_, vector<size_t> co
 	assert(types.size() == columns.size());
 	is_variable.resize(columns.size());
 	type_sizes.resize(columns.size());
-	for (size_t i = 0; i < columns.size(); i++) {
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		auto type = types[i];
 		type_sizes[i] = GetTypeIdSize(type);
 		if (TypeIsConstantSize(type)) {
@@ -48,7 +48,7 @@ void TupleSerializer::Serialize(DataChunk &chunk, Tuple targets[]) {
 	// first compute the sizes, if there are variable length columns
 	// if there are none, we can just use the base size
 	uint8_t *target_locations[STANDARD_VECTOR_SIZE];
-	for (size_t i = 0; i < chunk.size(); i++) {
+	for (uint64_t i = 0; i < chunk.size(); i++) {
 		targets[i].size = base_size;
 		targets[i].data = unique_ptr<uint8_t[]>(new uint8_t[targets[i].size]);
 		target_locations[i] = targets[i].data.get();
@@ -58,28 +58,28 @@ void TupleSerializer::Serialize(DataChunk &chunk, Tuple targets[]) {
 }
 
 void TupleSerializer::Serialize(DataChunk &chunk, uint8_t *targets[]) {
-	size_t offset = 0;
-	for (size_t i = 0; i < columns.size(); i++) {
+	uint64_t offset = 0;
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		SerializeColumn(chunk, targets, i, offset);
 	}
 }
 
 void TupleSerializer::Deserialize(Vector &source, DataChunk &chunk) {
-	for (size_t i = 0; i < columns.size(); i++) {
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		DeserializeColumn(source, i, chunk.data[columns[i]]);
 	}
 }
 
-void TupleSerializer::Serialize(vector<char *> &column_data, size_t offset, uint8_t *target) {
-	for (size_t i = 0; i < columns.size(); i++) {
+void TupleSerializer::Serialize(vector<char *> &column_data, uint64_t offset, uint8_t *target) {
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		auto source = column_data[columns[i]] + type_sizes[i] * offset;
 		memcpy(target, source, type_sizes[i]);
 		target += type_sizes[i];
 	}
 }
 
-void TupleSerializer::Deserialize(vector<char *> &column_data, size_t offset, uint8_t *target) {
-	for (size_t i = 0; i < columns.size(); i++) {
+void TupleSerializer::Deserialize(vector<char *> &column_data, uint64_t offset, uint8_t *target) {
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		auto source = column_data[columns[i]] + type_sizes[i] * offset;
 		memcpy(source, target, type_sizes[i]);
 		target += type_sizes[i];
@@ -88,9 +88,9 @@ void TupleSerializer::Deserialize(vector<char *> &column_data, size_t offset, ui
 
 void TupleSerializer::Serialize(vector<char *> &column_data, Vector &index_vector, uint8_t *targets[]) {
 	auto indices = (uint64_t *)index_vector.data;
-	size_t offset = 0;
-	for (size_t i = 0; i < columns.size(); i++) {
-		for (size_t j = 0; j < index_vector.count; j++) {
+	uint64_t offset = 0;
+	for (uint64_t i = 0; i < columns.size(); i++) {
+		for (uint64_t j = 0; j < index_vector.count; j++) {
 			auto index = index_vector.sel_vector ? indices[index_vector.sel_vector[j]] : indices[j];
 			auto source = column_data[columns[i]] + type_sizes[i] * index;
 			memcpy(targets[j] + offset, source, type_sizes[i]);
@@ -100,25 +100,25 @@ void TupleSerializer::Serialize(vector<char *> &column_data, Vector &index_vecto
 }
 
 void TupleSerializer::SerializeUpdate(vector<char *> &column_data, vector<column_t> &affected_columns,
-                                      DataChunk &update_chunk, Vector &index_vector, size_t index_offset,
+                                      DataChunk &update_chunk, Vector &index_vector, uint64_t index_offset,
                                       Tuple targets[]) {
 	auto indices = (uint64_t *)index_vector.data;
 	assert(index_vector.count == update_chunk.size());
 
 	// first initialize the tuples
-	for (size_t i = 0; i < index_vector.count; i++) {
+	for (uint64_t i = 0; i < index_vector.count; i++) {
 		targets[i].size = base_size;
 		targets[i].data = unique_ptr<uint8_t[]>(new uint8_t[targets[i].size]);
 	}
 
 	// now copy the data to the tuples
-	size_t offset = 0;
-	for (size_t i = 0; i < columns.size(); i++) {
+	uint64_t offset = 0;
+	for (uint64_t i = 0; i < columns.size(); i++) {
 		auto column = columns[i];
 		auto update_column = affected_columns[i];
 		if (update_column == (column_t)-1) {
 			// fetch from base column
-			for (size_t j = 0; j < index_vector.count; j++) {
+			for (uint64_t j = 0; j < index_vector.count; j++) {
 				auto index =
 				    (index_vector.sel_vector ? indices[index_vector.sel_vector[j]] : indices[j]) - index_offset;
 				auto source = column_data[column] + type_sizes[i] * index;
@@ -127,7 +127,7 @@ void TupleSerializer::SerializeUpdate(vector<char *> &column_data, vector<column
 		} else {
 			// fetch from update column
 			auto baseptr = update_chunk.data[update_column].data;
-			for (size_t j = 0; j < index_vector.count; j++) {
+			for (uint64_t j = 0; j < index_vector.count; j++) {
 				auto index = update_chunk.sel_vector ? update_chunk.sel_vector[j] : j;
 				auto source = baseptr + type_sizes[i] * index;
 				memcpy(targets[j].data.get() + offset, source, type_sizes[i]);
@@ -137,7 +137,8 @@ void TupleSerializer::SerializeUpdate(vector<char *> &column_data, vector<column
 	}
 }
 
-static void SerializeValue(uint8_t *target_data, Vector &col, size_t index, size_t result_index, size_t type_size) {
+static void SerializeValue(uint8_t *target_data, Vector &col, uint64_t index, uint64_t result_index,
+                           uint64_t type_size) {
 	if (col.nullmask[index]) {
 		SetNullValue(target_data, col.type);
 	} else {
@@ -145,18 +146,18 @@ static void SerializeValue(uint8_t *target_data, Vector &col, size_t index, size
 	}
 }
 
-void TupleSerializer::SerializeColumn(DataChunk &chunk, uint8_t *targets[], size_t column_index, size_t &offset) {
+void TupleSerializer::SerializeColumn(DataChunk &chunk, uint8_t *targets[], uint64_t column_index, uint64_t &offset) {
 	const auto column = columns[column_index];
 	const auto type_size = type_sizes[column_index];
 	assert(types[column_index] == chunk.data[column].type);
-	VectorOperations::Exec(chunk.data[column], [&](size_t i, size_t k) {
+	VectorOperations::Exec(chunk.data[column], [&](uint64_t i, uint64_t k) {
 		auto target_data = targets[k] + offset;
 		SerializeValue(target_data, chunk.data[column], i, k, type_size);
 	});
 	offset += type_size;
 }
 
-void TupleSerializer::DeserializeColumn(Vector &source, size_t column_index, Vector &target) {
+void TupleSerializer::DeserializeColumn(Vector &source, uint64_t column_index, Vector &target) {
 	const auto type_size = type_sizes[column_index];
 	assert(types[column_index] == target.type);
 	target.count = source.count;
@@ -175,7 +176,7 @@ int TupleSerializer::Compare(const uint8_t *a, const uint8_t *b) {
 		return memcmp(a, b, base_size);
 	} else {
 		int cmp = 0;
-		for (size_t i = 0; i < columns.size(); i++) {
+		for (uint64_t i = 0; i < columns.size(); i++) {
 			auto type_size = type_sizes[i];
 			if (is_variable[i]) {
 				auto left = *((const char **)a);
@@ -200,13 +201,13 @@ TupleComparer::TupleComparer(TupleSerializer &left, TupleSerializer &right) : le
 	left_offsets.resize(left.columns.size());
 	right_offsets.resize(left.columns.size());
 
-	size_t left_offset = 0;
-	for (size_t i = 0; i < left.columns.size(); i++) {
+	uint64_t left_offset = 0;
+	for (uint64_t i = 0; i < left.columns.size(); i++) {
 		left_offsets[i] = left_offset;
-		right_offsets[i] = (size_t)-1;
+		right_offsets[i] = (uint64_t)-1;
 
-		size_t right_offset = 0;
-		for (size_t j = 0; j < right.columns.size(); j++) {
+		uint64_t right_offset = 0;
+		for (uint64_t j = 0; j < right.columns.size(); j++) {
 			if (left.columns[i] == right.columns[j]) {
 				assert(left.types[i] == right.types[j]);
 
@@ -218,13 +219,13 @@ TupleComparer::TupleComparer(TupleSerializer &left, TupleSerializer &right) : le
 		}
 		// assert that we found the column
 		left_offset += left.type_sizes[i];
-		assert(right_offsets[i] != (size_t)-1);
+		assert(right_offsets[i] != (uint64_t)-1);
 	}
 }
 
 int TupleComparer::Compare(const uint8_t *left_data, const uint8_t *right_data) {
 	int cmp;
-	for (size_t i = 0; i < left_offsets.size(); i++) {
+	for (uint64_t i = 0; i < left_offsets.size(); i++) {
 		auto left_element = left_data + left_offsets[i];
 		auto right_element = right_data + right_offsets[i];
