+++ src/include/common/types/chunk_collection.hpp
@@ -8,6 +8,7 @@
 
 #pragma once
 
+#include "common/enums/order_type.hpp"
 #include "common/types/data_chunk.hpp"
 
 namespace duckdb {
@@ -26,14 +27,14 @@ public:
 	}
 
 	//! The total amount of elements in the collection
-	size_t count;
+	uint64_t count;
 	//! The set of data chunks in the collection
 	vector<unique_ptr<DataChunk>> chunks;
 	//! The types of the ChunkCollection
 	vector<TypeId> types;
 
 	//! The amount of columns in the ChunkCollection
-	size_t column_count() {
+	uint64_t column_count() {
 		return types.size();
 	}
 
@@ -41,11 +42,11 @@ public:
 	void Append(DataChunk &new_chunk);
 
 	//! Gets the value of the column at the specified index
-	Value GetValue(size_t column, size_t index);
+	Value GetValue(uint64_t column, uint64_t index);
 	//! Sets the value of the column at the specified index
-	void SetValue(size_t column, size_t index, Value value);
+	void SetValue(uint64_t column, uint64_t index, Value value);
 
-	vector<Value> GetRow(size_t index);
+	vector<Value> GetRow(uint64_t index);
 
 	string ToString() const {
 		return chunks.size() == 0 ? "ChunkCollection [ 0 ]"
@@ -54,7 +55,7 @@ public:
 	void Print();
 
 	//! Gets a reference to the chunk at the given index
-	DataChunk &GetChunk(size_t index) {
+	DataChunk &GetChunk(uint64_t index) {
 		return *chunks[LocateChunk(index)];
 	}
 
@@ -62,15 +63,15 @@ public:
 	//! Reorders the rows in the collection according to the given indices. NB: order is changed!
 	void Reorder(uint64_t order[]);
 
-	void MaterializeSortedChunk(DataChunk &target, uint64_t order[], size_t start_offset);
+	void MaterializeSortedChunk(DataChunk &target, uint64_t order[], uint64_t start_offset);
 
 	//! Returns true if the ChunkCollections are equivalent
 	bool Equals(ChunkCollection &other);
 
 private:
 	//! Locates the chunk that belongs to the specific index
-	size_t LocateChunk(size_t index) {
-		size_t result = index / STANDARD_VECTOR_SIZE;
+	uint64_t LocateChunk(uint64_t index) {
+		uint64_t result = index / STANDARD_VECTOR_SIZE;
 		assert(result < chunks.size());
 		return result;
 	}
