+++ src/include/planner/logical_operator.hpp
@@ -10,6 +10,7 @@
 
 #include "catalog/catalog.hpp"
 #include "common/common.hpp"
+#include "common/enums/logical_operator_type.hpp"
 #include "planner/expression.hpp"
 #include "planner/logical_operator_visitor.hpp"
 
@@ -47,26 +48,26 @@ public:
 	void ResolveOperatorTypes();
 
 	virtual string ParamsToString() const;
-	virtual string ToString(size_t depth = 0) const;
+	virtual string ToString(uint64_t depth = 0) const;
 	void Print();
 
 	void AddChild(unique_ptr<LogicalOperator> child) {
 		children.push_back(move(child));
 	}
 
-	virtual size_t EstimateCardinality() {
+	virtual uint64_t EstimateCardinality() {
 		// simple estimator, just take the max of the children
-		size_t max_cardinality = 0;
+		uint64_t max_cardinality = 0;
 		for (auto &child : children) {
 			max_cardinality = std::max(child->EstimateCardinality(), max_cardinality);
 		}
 		return max_cardinality;
 	}
 
-	virtual size_t ExpressionCount();
-	virtual Expression *GetExpression(size_t index);
+	virtual uint64_t ExpressionCount();
+	virtual Expression *GetExpression(uint64_t index);
 	virtual void ReplaceExpression(std::function<unique_ptr<Expression>(unique_ptr<Expression> expression)> callback,
-	                               size_t index);
+	                               uint64_t index);
 
 protected:
 	//! Resolve types for this specific operator
