+++ src/include/planner/operator/logical_order.hpp
@@ -22,10 +22,10 @@ public:
 
 	vector<BoundOrderByNode> orders;
 
-	size_t ExpressionCount() override;
-	Expression *GetExpression(size_t index) override;
+	uint64_t ExpressionCount() override;
+	Expression *GetExpression(uint64_t index) override;
 	void ReplaceExpression(std::function<unique_ptr<Expression>(unique_ptr<Expression> expression)> callback,
-	                       size_t index) override;
+	                       uint64_t index) override;
 
 protected:
 	void ResolveTypes() override {
