+++ src/include/storage/write_ahead_log.hpp
@@ -87,7 +87,7 @@ public:
 
 private:
 	template <class T> void Write(T val);
-	void WriteData(uint8_t *dataptr, size_t data_size);
+	void WriteData(uint8_t *dataptr, uint64_t data_size);
 
 	void WriteEntry(wal_type_t type, Serializer &serializer);
